requirements.txt
qdrant-client
voyageai
langchain
langchain-community
langchain-openai
streamlit
watchdog
python-dotenv
pydantic
pandas
